add near.state
