+++ package.json
@@ -11,6 +11,8 @@
     "@ant-design/icons": "^4.7.0",
     "@zeit/next-sass": "^1.0.1",
     "antd": "^4.16.13",
+    "mobx": "^6.3.6",
+    "mobx-react": "^7.2.1",
     "near-api-js": "^0.43.1",
     "near-seed-phrase": "^0.2.0",
     "next": "12.0.2",
